+++ helper/transaction_model.go
@@ -3,5 +3,5 @@ package helper
 type TransactionRequest struct {
 	Customer string
 	BookID   uint8
-	Quantity float32
+	Quantity uint8
 }
